refactor(jsx): remove dependencies to jsx/dom from jsx/index.ts
