+++ deno_dist/jsx/dom/index.ts
@@ -1,3 +1,5 @@
 export { render } from './render.ts'
 export { Suspense, ErrorBoundary } from './components.ts'
-export { createContext, useContext } from './context.ts'
+export { useContext } from '../context.ts'
+export type { Context } from '../context.ts'
+export { createContext } from './context.ts'
